+++ openid_provider/conf.py
@@ -1,14 +1,15 @@
 import os
-import tempfile
 from django.conf import settings
-
-tempdir = tempfile.gettempdir()
 
 STORE = getattr(settings, 'OPENID_PROVIDER_STORE',
                 'openid.store.filestore.FileOpenIDStore')
 
-FILESTORE_PATH = getattr(settings, 'OPENID_PROVIDER_FILESTORE_PATH',
-                         os.path.join(tempdir, 'openid-filestore'))
+if STORE == 'openid.store.filestore.FileOpenIDStore':
+    import tempfile
+    tempdir = tempfile.gettempdir()
+    
+    FILESTORE_PATH = getattr(settings, 'OPENID_PROVIDER_FILESTORE_PATH',
+                             os.path.join(tempdir, 'openid-filestore'))
 
 SREG_DATA_CALLBACK = getattr(settings, 'OPENID_PROVIDER_SREG_DATA_CALLBACK',
                              'openid_provider.utils.get_default_sreg_data')
@@ -17,3 +18,10 @@ AX_DATA_CALLBACK = getattr(settings, 'OPENID_PROVIDER_AX_DATA_CALLBACK',
                            'openid_provider.utils.get_default_ax_data')
 
 AX_EXTENSION = getattr(settings, 'OPENID_PROVIDER_AX_EXTENSION', False)
+
+AUTH_USER_MODEL = getattr(settings, 'AUTH_USER_MODEL', 'auth.User')
+
+# RPs without relying party verification mechanisms will be each time
+# redirected to decide page, set to True to disable this:
+FAILED_DISCOVERY_AS_VALID = getattr(
+    settings, 'OPENID_FAILED_DISCOVERY_AS_VALID', False)